Specify and enforce node and npm requirements
Since this project uses workspaces it requires npm 7 or greater. This
can take a minute to figure out if you don't know it. This change causes
you to set npm to the appropriate version if you try to use it, e.g.
when installing with `npm install`

I specified node>=16 simply because npm seemed to think that was the
required version when specifying npm>=7.